% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_colors.R
\docType{data}
\name{bird_colors}
\alias{bird_colors}
\title{Climatic drivers of coloration in Furnariidae species}
\format{
An object of class data.frame
}
\usage{
data(bird_colors)
}
\description{
Data for redness and brightness of different plumage patches in Furnariidae species
and associated climatic data
}
\references{
Marcondes et al. (2021) The American Naturalist 197:592-606
(\href{https://pubmed.ncbi.nlm.nih.gov/33908827/}{PubMed})

Seeholzer et al. (2017) Evolution 71:702:715
(\href{https://pubmed.ncbi.nlm.nih.gov/28071791/}{PubMed})
}
\keyword{datasets}
