% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_thoracicus.R
\docType{data}
\name{data_thoracicus}
\alias{data_thoracicus}
\title{Diet and morphology data for Microlophus thoracicus (Squamata: Tropiduridae)}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(data_thoracicus)
}
\description{
Data for head morphology and proportion of plant material in the diet
of the lizard Microlophus thoracicus.
}
\references{
Toyama et al. (2018) Genetics 195:1077-1086
(\href{https://pubmed.ncbi.nlm.nih.gov/30170747/}{PubMed})
}
\keyword{datasets}
