% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jnt_cat.R
\name{jnt_cat}
\alias{jnt_cat}
\title{Calculation and visualization of regions of non-significance to assess the
influence of categorical moderators}
\usage{
jnt_cat(
  X,
  Y,
  g,
  data,
  plot.full = F,
  phylo = F,
  tree,
  cols = c("black", "black"),
  sym = c(16, 1)
)
}
\arguments{
\item{X}{A character string defining the name of the covariate (e.g., size in
an allometry analysis). Must be the same as the name of the variable in the
dataset (see argument “data”).}

\item{g}{A character string defining the name of a categorical moderator
(e.g., males and females, herbivorous and carnivorous, etc). Must be the same
as the name of the variable in the dataset (see argument “data”). The
variable must have two levels.}

\item{data}{The dataset.}

\item{phylo}{A logical. It indicates whether a phylogenetically-informed analysis
will be performed (i.e., PGLS). “F” by default.}

\item{tree}{A “phylo” class object. A phylogeny that must match the species in the data.}

\item{cols}{A vector of strings defining the symbol colors to be used in the
plot. By default, c(“black”, “black”) is used, which combines with the default
in the argument “sym” to present two groups of datapoints as open and close.}

\item{sym}{A vector of strings defining the symbols to be used to represent
distinct groups in the plot. Use same symbol codes as in the argument “pch”
in the R base function “plot”. By default, “c(16, 1)” is used, which combines
with the default in the argument “cols” to present two groups of datapoints
as open and close.}

\item{plot.lim}{A logical. It indicates whether the plot should show the
JN non-significance regions even if they don’t or just slightly overlap the data.
The default option is “F”, meaning that the plot limits will depend only on the data.}
}
\description{
Produces a plot showing which values of a categorical moderator have a
significant influence on the relationship between the dependent and
independent variables.
}
