% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jnt_cont.R
\name{jnt_cont}
\alias{jnt_cont}
\title{Calculation and visualization of regions of non-significance to assess the
influence of continuous moderators}
\usage{
jnt_cont(X, Y, g, data, phylo = F, tree, res = 100)
}
\arguments{
\item{X}{A character string defining the name of the covariate (e.g., size in
an allometry analysis). Must be the same as the name of the variable in the
dataset (see argument “data”).}

\item{Y}{A character string defining the name of the dependent variable. Must
be the same as the name of the variable in the dataset (see argument “data”).}

\item{g}{A character string defining the name of a continuous moderator (e.g., …, etc).
Must be the same as the name of the variable in the dataset (see argument “data”).
The variable must be continuous.}

\item{data}{The dataset.}

\item{phylo}{A logical. It indicates whether a phylogenetically-informed
analysis will be performed (i.e., PGLS). “F” by default.}

\item{tree}{A “phylo” class object. A phylogeny that must match the species
in the data.}

\item{res}{A numerical value that aids in the plotting of regions of (non)significance.
Default=100, higher numbers increase the intensity of colors}
}
\description{
Produces a plot showing which values of a continuous moderator have a
significant influence on the relationship between the dependent and
independent variables.
}
