% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizard_home_range.R
\docType{data}
\name{lizard_home_range}
\alias{lizard_home_range}
\title{Determinants of home range size in sleepy lizards}
\format{
An object of class data.frame
}
\usage{
data(lizard_home_range)
}
\description{
Dataset includes data on the determinants of home range size in the lizard
Tiliqua rugosa
}
\references{
Payne et al. (2021) Ecological Monographs 92:e1519
(\doi{10.1002/ecm.1519})
}
\keyword{datasets}
