% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microlophus.R
\docType{data}
\name{microlophus}
\alias{microlophus}
\title{Morphology of Microlophus lizards (Squamata: Tropiduridae)}
\format{
An object of class data.frame
}
\usage{
data(microlophus)
}
\description{
Data for body size and head morphology of Microlophus lizards
}
\references{
Toyama et al. (2018) Zoology 129:45-53
(\href{https://pubmed.ncbi.nlm.nih.gov/30170747/}{PubMed})
}
\keyword{datasets}
