% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_Furnariidae.R
\docType{phylogeny}
\name{tree_Furnariidae}
\alias{tree_Furnariidae}
\title{Phylogeny of Furnariidae}
\format{
An object of class phylo
}
\usage{
data(tree_F)
}
\description{
Phylogeny of Furnariidae
}
\references{
Marcondes et al. (2021) The American Naturalist 197:592-606
(\href{https://pubmed.ncbi.nlm.nih.gov/33908827/}{PubMed})
}
\keyword{phylogeny}
