% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jnt_cat.R
\name{jnt_cat}
\alias{jnt_cat}
\title{Calculation and visualization of regions of non-significance to assess the
influence of categorical moderators}
\usage{
jnt_cat(
  X,
  Y,
  m,
  data,
  plot.full = F,
  phylo = F,
  correlation,
  cols = c("black", "black"),
  pch = c(16, 1),
  cex = 1,
  xlab = X,
  ylab = Y,
  legend = F,
  lty = c(1, 2),
  line.col = c("black", "black"),
  lwd = c(1, 1)
)
}
\arguments{
\item{X}{A character string defining the name of the covariate (e.g., size in
an allometry analysis). Must be the same as the name of the variable in the
dataset (see argument “data”).}

\item{Y}{A character string defining the name of the dependent variable. Must
be the same as the name of the variable in the dataset (see argument “data”).}

\item{m}{A character string defining the name of a categorical moderator
(e.g., males and females, herbivorous and carnivorous, etc). Must be the same
as the name of the variable in the dataset (see argument “data”). The
variable must have two levels.}

\item{data}{The dataset.}

\item{plot.full}{A logical. It indicates whether the plot should show the
JN non-significance regions even if they don’t or just slightly overlap the data.
The default option is “F”, meaning that the plot limits will depend only on the data.}

\item{phylo}{A logical. It indicates whether a phylogenetically-informed analysis
will be performed (i.e., PGLS). “F” by default.}

\item{correlation}{an optional \link{corStruct} object describing the within-group
correlation structure. See the documentation of \link{corClasses} for a description of
the available corStruct classes. If a grouping variable is to be used, it must be
specified in the form argument to the corStruct constructor. Defaults to NULL,
corresponding to uncorrelated errors.}

\item{cols}{A vector of strings defining the symbol colors to be used in the
plot. By default, c(“black”, “black”) is used, which combines with the default
in the argument “pch” to present two groups of datapoints as open and close.}

\item{pch}{A vector of strings defining the symbols to be used to represent
distinct groups in the plot. Use same symbol codes as in the argument “pch”
in the R base function “plot”. By default, “c(16, 1)” is used, which combines
with the default in the argument “cols” to present two groups of datapoints
as open and close.}
}
\description{
Produces a plot showing which values of a categorical moderator have a
significant influence on the relationship between the dependent and
independent variables.
}
