% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jnt_cont.R
\name{jnt_cont}
\alias{jnt_cont}
\title{Calculation and visualization of regions of non-significance to assess the
influence of continuous moderators}
\usage{
jnt_cont(
  X,
  Y,
  g,
  data,
  phylo = F,
  correlation,
  res = 100,
  xlab = X,
  ylab = Y,
  sig_color = "red",
  nonsig_color = "grey",
  col.gradient = F,
  max_col_grad = "red",
  min_col_grad = "blue"
)
}
\arguments{
\item{X}{A character string defining the name of the covariate (e.g., size in
an allometry analysis). Must be the same as the name of the variable in the
dataset (see argument “data”).}

\item{Y}{A character string defining the name of the dependent variable. Must
be the same as the name of the variable in the dataset (see argument “data”).}

\item{g}{A character string defining the name of a continuous moderator (e.g., …, etc).
Must be the same as the name of the variable in the dataset (see argument “data”).
The variable must be continuous.}

\item{data}{The dataset.}

\item{phylo}{A logical. It indicates whether a phylogenetically-informed
analysis will be performed (i.e., PGLS). “F” by default.}

\item{correlation}{an optional \link{corStruct} object describing the within-group
correlation structure. See the documentation of corClasses for a description of
the available corStruct classes. If a grouping variable is to be used, it must be
specified in the form argument to the corStruct constructor. Defaults to NULL,
corresponding to uncorrelated errors.}

\item{res}{A numerical value that aids in the plotting of regions of (non)significance.
Default=100, higher numbers increase the intensity of colors}
}
\description{
Produces a plot showing which values of a continuous moderator have a
significant influence on the relationship between the dependent and
independent variables.
}
