% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_Furnariidae.R
\docType{data}
\name{tree_Furnariidae}
\alias{tree_Furnariidae}
\title{Phylogeny of Furnariidae}
\format{
An object of class phylo
}
\usage{
data(tree_Furnariidae)
}
\description{
Phylogeny of Furnariidae
}
\references{
Harvey et al. (2020) Science 370:1343-1348
(\href{https://pubmed.ncbi.nlm.nih.gov/33303617/}{PubMed})
}
\keyword{phylogeny}
